All fileIO methods are fully implemented
fileLoad method is implemented by a left and a right.
fileSave and fileSaveAs methods are also implemented by a left and a right, individually.
fileSave and fileSaveAs methods are not completly tested.
the test for fileSave and fileSaveAs especially for the blocks which is made when the compare or merge is done, must be conducted first.